move input tests under input package
